Make sure that resources are available to NQP when compiling the debugger
Currently, NQP needs to access the Perl6::Grammar class from Rakudo,
as well as some of the dynext ops from Rakudo.  If using
make install DESTDIR=$dir, they've been installed to a staging location
rather than their final destination, so we need to show NQP where they
are.  We do this by creating symbolic links to the required resources.

PLEASE, PLEASE, PLEASE fix this up if NQP adds a way to control where
it looks for Perl 6 modules or for dynext files!  This is a horrible
hack and I'm not terribly proud of it!