Update the module test method
A conservative PATH is now set, with the bin directories from Rakudo
Star added on to it. This should ensure that the right Raku is being
used.

Additionally, any failing modules will be reported at the end, and the
exitcode of the test action becomes non-zero if any module failed their
tests.