Introduce a GitHub Workflow to build a binary packages
This workflow will specifically build a package for amd64 machines using
a GNU libc, but allows for extending into other territories later on.
Using this, we *should* be able to automatically build for Windows and
Mac users too, solving that problem forever if someone wants to spend
some time setting it up.

Once other architectures for Linux-based systems gets introduced on
GitHub Actions, these could be build in a similar manner too.